ec2.securitygroup: Avoid late init if it seems unsafe
If the structure of the ingress/egress rules are different, or any port
numbers/protocols do not match, avoid late init because we will not be
able to do it safely. The array does not have a unique key.

Also avoid late init of description because I have not seen that AWS
sets description.

Updates #618

Signed-off-by: Carl Henrik Lunde <chlunde@ifi.uio.no>