rds.rdsinstance: autoscaled DB is up to date
Make rds.CreatePatch exclude AllocatedStorage if autoscaling is enabled,
and the requested AllocatedStorage is lower than was is currently
allocated.

This allows a user to manually scale *up* storage even if autoscaling is
enabled, and prevents crossplane from attempting to scale *down* if
AWS has performed autoscaling.

One problem is that this change silently ignores any attempts to manually
downscale when autoscaling is enabled. Since downscaling is not
supported and it is not possible to distinguish this from autoscaling, I
don't see any other solutions.

Before this change, the new tests would fail:

    --- FAIL: TestObserve (0.00s)
        --- FAIL: TestObserve/AutoscaledStorageIsUpToDate (0.00s)
            rdsinstance_test.go:354: r: -want, +got:
                  managed.ExternalObservation{
                  	ResourceExists:          true,
                - 	ResourceUpToDate:        true,
                + 	ResourceUpToDate:        false,
                  	ResourceLateInitialized: false,
                  	ConnectionDetails:       nil,
                  	Diff:                    "",
                  }
    --- FAIL: TestUpdate (0.00s)
        --- FAIL: TestUpdate/AutoscaleExcludeStorage (0.00s)
            rdsinstance_test.go:674: r: -want, +got:
                  interface{}(
                + 	e"cannot modify RDS instance: AllocatedStorage must not be set when on a modify request when AWS has autoscaled the storage",
                  )

Signed-off-by: Carl Henrik Lunde <chlunde@ifi.uio.no>