iam.iamrole: Use improved policy comparision function
Reuse the improved code for IAM policy comparision from the ECR package.

Currently two IAM assume role policy documents do not match when they
use maps in condition keys, because JSON-encoded Go maps are not ordered.

This has already been handled in the ECR package, so we can reuse this
code. It should probably be moved to a common package.

Add a test to trigger a diff in a policy, and also verify that the error
code from diff is as expected.

Fixes #792
Updates #704

Signed-off-by: Carl Henrik Lunde <chlunde@ifi.uio.no>